Getting there with backtesting. Implementation of strategy next

+++ domain/Candle.cs
@@ -2,7 +2,7 @@ namespace coinbase_bot.domain;
 
 public class Candle
 {
-    public decimal Start { get; set; }
+    public long Start { get; set; }
     public decimal Low { get; set; }
     public decimal High { get; set; }
     public decimal Open { get; set; }
